feat: ego server stop|restart now use REST /service/admin/down endpoint
By default, attempting to stop the server will use a REST call to tell
the server to do a clean shutdown. This closes the log file cleanly,
etc.

Note that a ego server restart _cannot_ be done this way if the
server in the profile is not the current host -- you can stop a
remote server (if you have admin privilege) but you cannot start
a remote server, so restart is not allowed.

If you want/need to stop or restart the server on the local machine
without using the REST interface to stop the server, use the "--force"
flag which instead uses direct OS calls to kill the server based on
the PID number of the server. This does not require admin privilege
(but does require you to be the owner of the server process or have
root privilege).

+++ commands/restart.go
@@ -2,10 +2,12 @@ package commands
 
 import (
 	"os"
+	"strings"
 	"time"
 
 	"github.com/google/uuid"
 	"github.com/tucats/ego/app-cli/cli"
+	"github.com/tucats/ego/app-cli/settings"
 	"github.com/tucats/ego/app-cli/ui"
 	"github.com/tucats/ego/defs"
 	"github.com/tucats/ego/errors"
@@ -22,73 +24,73 @@ func Restart(c *cli.Context) error {
 	}
 
 	serverStatus, err := killExistingServer(c)
-	if err == nil {
-		args := serverStatus.Args
-
-		// Set up the new ID. If there was one already (because this might be
-		// a restart operation) then update the UUID value. If not, add the uuid
-		// command line option.
-		logID := uuid.New()
-		found := false
-
-		for i, v := range args {
-			if v == "--session-uuid" {
-				args[i+1] = logID.String()
-				found = true
-
-				break
-			}
-		}
-
-		if !found {
-			args = append(args, "--session-uuid", logID.String())
-		}
-
-		if c.Boolean("new-token") {
-			args = append(args, "--new-token")
-		}
-
-		// Sleep for one second. This guarantees that the log file stamp of the new log
-		// will not be the same as the old log stamp.
-		time.Sleep(1 * time.Second)
-
-		// Launch the new process
-		pid, err := fork.Run(args[0], args)
-		if err == nil {
-			serverStatus.PID = pid
-			serverStatus.ID = logID.String()
-
-			// Scan over args and remove any instance of "--new-token". These are
-			// not saved in the pid file, so this option is only a "one-shot"
-			for i, v := range args {
-				if v == "--new-token" {
-					args = append(args[:i], args[i+1:]...)
-				}
-			}
-
-			// Write the new status to the pid file.
-			// We need to write it again, because the log file name might have changed.
-			// Note that the log file name is not included in the status.Args slice.
-			serverStatus.Args = args
-			err = server.WritePidFile(c, *serverStatus)
-
-			if ui.OutputFormat == ui.TextFormat {
-				ui.Say("msg.server.started", map[string]interface{}{
-					"pid": pid,
-				})
-			} else {
-				serverState, _ := server.ReadPidFile(c)
-				_ = commandOutput(serverState)
-			}
-		} else {
-			_ = server.RemovePidFile(c)
-		}
-
-		if err != nil {
-			err = errors.New(err)
-		}
-
+	if !errors.Nil(err) {
 		return err
+	}
+
+	args := serverStatus.Args
+
+	// Set up the new ID. If there was one already (because this might be
+	// a restart operation) then update the UUID value. If not, add the uuid
+	// command line option.
+	logID := uuid.New()
+	found := false
+
+	for i, v := range args {
+		if v == "--session-uuid" {
+			args[i+1] = logID.String()
+			found = true
+
+			break
+		}
+	}
+
+	if !found {
+		args = append(args, "--session-uuid", logID.String())
+	}
+
+	if c.Boolean("new-token") {
+		args = append(args, "--new-token")
+	}
+
+	// Sleep for one second. This guarantees that the log file stamp of the new log
+	// will not be the same as the old log stamp.
+	time.Sleep(1 * time.Second)
+
+	// Launch the new process
+	pid, err := fork.Run(args[0], args)
+	if err == nil {
+		serverStatus.PID = pid
+		serverStatus.ID = logID.String()
+
+		// Scan over args and remove any instance of "--new-token". These are
+		// not saved in the pid file, so this option is only a "one-shot"
+		for i, v := range args {
+			if v == "--new-token" {
+				args = append(args[:i], args[i+1:]...)
+			}
+		}
+
+		// Write the new status to the pid file.
+		// We need to write it again, because the log file name might have changed.
+		// Note that the log file name is not included in the status.Args slice.
+		serverStatus.Args = args
+		err = server.WritePidFile(c, *serverStatus)
+
+		if ui.OutputFormat == ui.TextFormat {
+			ui.Say("msg.server.started", map[string]interface{}{
+				"pid": pid,
+			})
+		} else {
+			serverState, _ := server.ReadPidFile(c)
+			_ = commandOutput(serverState)
+		}
+	} else {
+		_ = server.RemovePidFile(c)
+	}
+
+	if err != nil {
+		err = errors.New(err)
 	}
 
 	return err
@@ -98,25 +100,54 @@ func Restart(c *cli.Context) error {
 // of the server if it was running, and an error code indicating if the server
 // was killed. If the server was not running, returns nil and no error.
 func killExistingServer(c *cli.Context) (*defs.ServerStatus, error) {
-	status, err := server.ReadPidFile(c)
-	if err == nil {
-		proc, e2 := os.FindProcess(status.PID)
-		if e2 == nil {
-			e2 = proc.Kill()
-			// If successful, and in text mode, report the stop to the console.
-			if e2 == nil && ui.OutputFormat == ui.TextFormat {
-				ui.Say("msg.server.stopped", map[string]interface{}{
-					"pid": status.PID,
-				})
+	if c.Boolean("force") {
+		status, err := server.ReadPidFile(c)
+		if err == nil {
+			proc, e2 := os.FindProcess(status.PID)
+			if e2 == nil {
+				e2 = proc.Kill()
+				// If successful, and in text mode, report the stop to the console.
+				if e2 == nil && ui.OutputFormat == ui.TextFormat {
+					ui.Say("msg.server.stopped", map[string]interface{}{
+						"pid": status.PID,
+					})
+				}
+			}
+
+			if e2 != nil {
+				err = errors.New(e2)
 			}
 		}
 
-		if e2 != nil {
-			err = errors.New(e2)
-		}
+		_ = server.RemovePidFile(c)
+
+		return status, err
 	}
 
-	_ = server.RemovePidFile(c)
+	// Not a force, let's try it the polite way. The current server (app.server
+	// or logon.server) must be the current server or this operation is invalid
+	// (cannot do a restart on another server)
+
+	hostname, err := os.Hostname()
+	if err != nil {
+		return nil, errors.New(err)
+	}
+
+	server := settings.Get(defs.ApplicationServerSetting)
+	if server == "" {
+		server = settings.Get(defs.LogonServerSetting)
+	}
+
+	server = strings.TrimPrefix(server, "http://")
+	server = strings.TrimPrefix(server, "https://")
+
+	server, _, _ = strings.Cut(server, ":")
+	if strings.EqualFold(server, hostname) {
+		return nil, errors.ErrNotLocalServer.Context(server)
+	}
+
+	// All good, let's try to send a REST stop command to the local server.
+	status, err := politeStop(c)
 
 	return status, err
 }
